+++ internal/api/router.go
@@ -33,6 +33,7 @@ func RegisterRouter(e *echo.Echo, repo port.Repository, cfg *config.Config, mtxC
 	handler.RegisterAsset(api, d)
 	handler.RegisterOperator(api, d)
 	handler.RegisterWorkflow(api, d)
+	handler.RegisterTask(api, d)
 	handler.RegisterAlgorithm(api, d)
 	handler.RegisterAlgorithmBinding(api, d)
 	handler.RegisterRecord(api, d)